refactor RulesBasedSamplerConfig
