+++ config/sampler_config.go
@@ -1,5 +1,10 @@
 package config
 
+import (
+	"fmt"
+	"github.com/honeycombio/refinery/sample"
+)
+
 type DeterministicSamplerConfig struct {
 	SampleRate int `validate:"required,gte=1"`
 }
@@ -35,4 +40,40 @@ type TotalThroughputSamplerConfig struct {
 	UseTraceLength               bool
 	AddSampleRateKeyToTrace      bool
 	AddSampleRateKeyToTraceField string `validate:"required_with=AddSampleRateKeyToTrace"`
+}
+
+type RulesBasedSamplerCondition struct {
+	Field    string
+	Operator string
+	Value    interface{}
+}
+
+func (r *RulesBasedSamplerCondition) String() string {
+	return fmt.Sprintf("%+v", *r)
+}
+
+type RulesBasedDownstreamSampler struct {
+	DynamicSampler    *DynamicSamplerConfig
+	EMADynamicSampler *EMADynamicSamplerConfig
+}
+
+type RulesBasedSamplerRule struct {
+	Name              string
+	SampleRate        int
+	Downstream        *RulesBasedDownstreamSampler
+	DownstreamSampler sample.Sampler
+	Drop              bool
+	Condition         []*RulesBasedSamplerCondition
+}
+
+func (r *RulesBasedSamplerRule) String() string {
+	return fmt.Sprintf("%+v", *r)
+}
+
+type RulesBasedSamplerConfig struct {
+	Rule []*RulesBasedSamplerRule
+}
+
+func (r *RulesBasedSamplerConfig) String() string {
+	return fmt.Sprintf("%+v", *r)
 }
